+++ src/node-control/service/src/http/http_server_task.rs
@@ -904,7 +904,7 @@ impl utoipa::Modify for BearerAuthAddon {
         common::app_config::BindingStatus,
         common::app_config::LogRotation,
         common::app_config::LogOutput,
-        common::app_config::TonCoreDeployLayout,
+        common::app_config::TonCoreDeployMode,
         super::config_handlers::ContractsAutomationSettingsResponse,
         super::config_handlers::ContractsAutomationSettingsUpdateRequest,
         super::config_handlers::ElectionsSettingsUpdateRequest,